Move IndexSourceNode to the SPI
Move IndexSourceNode to the SPI so it can be used in
ConnectorPlanOptimizers and PlanCheckers
