+++ presto-main/src/main/java/com/facebook/presto/sql/planner/plan/InternalPlanVisitor.java
@@ -37,11 +37,6 @@ public abstract class InternalPlanVisitor<R, C>
         return visitPlan(node, context);
     }
 
-    public R visitIndexSource(IndexSourceNode node, C context)
-    {
-        return visitPlan(node, context);
-    }
-
     public R visitIndexJoin(IndexJoinNode node, C context)
     {
         return visitPlan(node, context);
