+++ presto-main/src/test/java/com/facebook/presto/sql/planner/assertions/IndexSourceMatcher.java
@@ -18,8 +18,8 @@ import com.facebook.presto.common.predicate.Domain;
 import com.facebook.presto.cost.StatsProvider;
 import com.facebook.presto.metadata.Metadata;
 import com.facebook.presto.spi.TableMetadata;
+import com.facebook.presto.spi.plan.IndexSourceNode;
 import com.facebook.presto.spi.plan.PlanNode;
-import com.facebook.presto.sql.planner.plan.IndexSourceNode;
 import com.google.common.collect.ImmutableMap;
 
 import java.util.Map;
