+++ presto-spi/src/main/java/com/facebook/presto/spi/plan/PlanVisitor.java
@@ -144,4 +144,9 @@ public abstract class PlanVisitor<R, C>
     {
         return visitPlan(node, context);
     }
+
+    public R visitIndexSource(IndexSourceNode node, C context)
+    {
+        return visitPlan(node, context);
+    }
 }
